Justin/feature/added new methods for meter packages

+++ src/main/java/io/github/lordship/meterbills/MeterBills.java
@@ -2,6 +2,7 @@ package io.github.lordship.meterbills;
 
 import io.github.lordship.meters.MeterMeasurement;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.OffsetDateTime;
 import java.util.UUID;
@@ -9,8 +10,8 @@ import java.util.UUID;
 public record MeterBills(
         UUID uuid,
         UUID billedMeter,
-        Integer billedAmount,
-        Integer rateAmount,
+        BigDecimal billedAmount, // BigDecimal should be used for currency
+        BigDecimal rateAmount,
         MeterMeasurement rateUnit,
         LocalDate periodStart,
         LocalDate periodEnd,
